pyproject.toml
[project]
name = "chess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.1.2",
    "openenv-core[core]>=0.2.3",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
]
